Minecraft is installing now! Minecraft can be installed to an instance by way of the MinecraftInstaller class. In general, these are the steps to follow: 1. Create an instance: Specify where the instance directory is, what version of Minecraft you want, and what side this instance should be (client or server). 2. Create a MinecraftInstaller object and pass the instance to its constructor. 3. Call .install() on the MinecraftInstaller object. It will then download all of the necessary resources and install them in the correct locations.
Currently, version info files and asset indices are not automatically downloaded. They will be in a future update.

JavaLauncher can now launch with either the name of a .jar file (via .setJarName()) or a classname (via .setClassName()). If both are set at launch time, an exception is generated.

Added validation to several classes that are deserialized from JSON. You can call .validate() on a deserialized object to ensure that valid values have been provided for each field.
Added .getPath() to AssetIndex.AssetObject. This returns a location for the asset based on its hash.
Added .shouldExtract() to ExtractRules. This returns true if the given path within the library's .jar file is permitted to be extracted, or false otherwise.
CompleteMinecraftVersion now parses the "assets" field as well (setAssets() / getAssets() ).
CompleteMinecraftVersion now has .hasClient(), .hasServer(), and .hasWindowsServer().
Fixed a bug in Library.isCompatible(). If compatibility rules were given, it considered the library compatible by default. The Minecraft Launcher specifies that the library should be incompatible by default in that case.
Changed Library.getPath() to Library.getArtifact(). This returns the specific artifact that should be downloaded for the system (which contains the URL, filesystem path, and SHA-1 hash) for the library and target.

Added MinecraftConstants to store constants shared by both MinecraftLauncher and MinecraftInstaller. These may eventually wind up somewhere else.
Added .getAssetIndex( id ) and .loadAssetIndex( id ) to MinecraftVersions to grab asset indices.