Fixed a bug in ForgeInstaller where the download of the Forge binary to a temporary file failed with a FileAlreadyExistsException If an exception occurs while downloading the Forge binary, the temporary file will now be deleted.
Added IO.copyAndSHA1( in, out, hash ) which is identical to IO.copy( in, out ) with one exception:
If hash is non-null, computes a SHA1 hash of the data as it passes between in and out.
If the computed hash doesn't match, a RuntimeException is thrown.

MinecraftInstaller now streams downloads to disk using IO.copyAndSHA1(), eliminating the need to read the whole file into memory before writing it to disk.

Added Misc.newSHA1() and Misc.newMD5() to return new instances of SHA-1 and MD5 MessageDigest objects, respectively, without the need to handle a checked exception.